Added parent class for buffs
See Bulbasaur.cs code for example
