+++ Pokemon/FirstGeneration/Normal/Bulbasaur/BulbasaurBuff.cs
@@ -4,35 +4,15 @@ using Terraria.ModLoader;
 
 namespace Terramon.Pokemon.FirstGeneration.Normal.Bulbasaur
 {
-    public class BulbasaurBuff : ModBuff
+    public class BulbasaurBuff : PokemonBuff
     {
+        public override string ProjectileName { get; } = nameof(Bulbasaur);
+
         public override void SetDefaults()
         {
+            base.SetDefaults();
             DisplayName.SetDefault("Bulbasaur");
             Description.SetDefault("A Bulbasaur is following you around!");
-            Main.buffNoTimeDisplay[Type] = true;
-            Main.vanityPet[Type] = true;
-        }
-
-        public override void Update(Player player, ref int buffIndex)
-        {
-            player.noFallDmg = true;
-
-            player.buffTime[buffIndex] = 40000;
-            TerramonPlayer modPlayer = (TerramonPlayer)player.GetModPlayer(mod, "TerramonPlayer");
-            modPlayer.bulbasaurPet = true;
-            bool petProjectileNotSpawned = true;
-            if (player.ownedProjectileCounts[mod.ProjectileType("Bulbasaur")] > 0)
-            {
-                petProjectileNotSpawned = false;
-            }
-
-
-            if (petProjectileNotSpawned && player.whoAmI == Main.myPlayer)
-            {
-                Projectile.NewProjectile(player.position.X + player.width / 2, player.position.Y + player.height / 2, 0f, 0f, mod.ProjectileType("Bulbasaur"), 0, 0f, player.whoAmI, 0f, 0f);
-            }
-
         }
     }
 }
